+++ tests/relay/helpers.ts
@@ -32,10 +32,6 @@ export interface RelayInternals {
 
 export const asRelayInternals = (relay: Relay): RelayInternals => relay as unknown as RelayInternals;
 
-export function assertExists<T>(value: T | null | undefined): asserts value is T {
-  expect(value).to.exist;
-}
-
 // Randomly generated key
 const defaultPrivateKey = '8841e004c6f47af679c91d9282adc62aeb9fabd19cdff6a9da5a358d0613c30a';
