+++ tests/server/acceptance/rpc_batch1.spec.ts
@@ -13,7 +13,8 @@ import Constants from '../../../src/relay/lib/constants';
 import { predefined } from '../../../src/relay/lib/errors/JsonRpcError';
 import { type TxPoolTransactionsByNonce } from '../../../src/relay/lib/txpool';
 import { BLOCK_NUMBER_ERROR, HASH_ERROR } from '../../../src/relay/lib/validators/constants';
-import { assertExists, overrideEnvsInMochaDescribe, withOverriddenEnvsInMochaTest } from '../../relay/helpers';
+import { assertExists } from '../../helpers/typeAssertions';
+import { overrideEnvsInMochaDescribe, withOverriddenEnvsInMochaTest } from '../../relay/helpers';
 import type MirrorClient from '../clients/mirrorClient';
 import type RelayClient from '../clients/relayClient';
 import type ServicesClient from '../clients/servicesClient';
